- Change navigation bar color - Add custom status bar background - The status bar text color (style lightContent) - Restructure the project to MVC - Add extension to UIColor.
