+++ YoutubeClone/Extension.swift
@@ -1,0 +1,30 @@
+//
+//  Extension.swift
+//  YoutubeClone
+//
+//  Created by Ky Nguyen on 10/14/16.
+//  Copyright © 2016 Ky Nguyen. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    
+    static func rgb(red: CGFloat, green: CGFloat, blue: CGFloat) -> UIColor {
+        return UIColor(red: red/255, green: green/255, blue: blue/255, alpha: 1)
+    }
+}
+
+extension UIView {
+    
+    func addConstraints(withFormat format: String, views: UIView...) {
+        var viewsDictionary = [String: UIView]()
+        for (index, view) in views.enumerated() {
+            let key = "v\(index)"
+            view.translatesAutoresizingMaskIntoConstraints = false
+            viewsDictionary[key] = view
+        }
+        
+        addConstraints(NSLayoutConstraint.constraints(withVisualFormat: format, options: NSLayoutFormatOptions(), metrics: nil, views: viewsDictionary))
+    }
+}
