+++ YoutubeClone/HomeViewController.swift
@@ -13,7 +13,13 @@ class HomeViewController: UICollectionViewController, UICollectionViewDelegateFl
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        navigationItem.title = "Home"
+        navigationController?.navigationBar.isTranslucent = true
+        let titleLabel = UILabel(frame: CGRect(x: 0, y: 0, width: view.frame.width, height: view.frame.height))
+        titleLabel.text = "Home"
+        titleLabel.textColor = UIColor.white
+        titleLabel.font = UIFont.systemFont(ofSize: 20)
+        navigationItem.titleView = titleLabel
+        
         collectionView?.backgroundColor = UIColor.white
         
         collectionView?.register(VideoCell.self, forCellWithReuseIdentifier: "cellId")
@@ -29,112 +35,12 @@ class HomeViewController: UICollectionViewController, UICollectionViewDelegateFl
     }
 
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: view.frame.width, height: 200)
+        let height = (view.frame.width - 32) * 9 / 16
+        return CGSize(width: view.frame.width, height: height + 84)
     }
     
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
         return 0
     }
     
-}
-
-class VideoCell: UICollectionViewCell {
-    
-    let thumbnailImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.backgroundColor = UIColor.blue
-        imageView.image = UIImage(named: "swift-banner")
-        imageView.contentMode = UIViewContentMode.scaleAspectFill
-        return imageView
-    }()
-    
-    let separatorView: UIView = {
-        let view = UIView()
-        view.backgroundColor = UIColor.black
-        return view
-    }()
-    
-    let userProfileImageView: UIImageView = {
-       let imageView = UIImageView()
-        imageView.backgroundColor = UIColor.green
-        return imageView
-    }()
-    
-    let titleLabel: UILabel = {
-       let label = UILabel()
-        label.backgroundColor = UIColor.purple
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let subtitleTextView: UITextView = {
-        let textView = UITextView()
-        textView.backgroundColor = UIColor.red
-        textView.translatesAutoresizingMaskIntoConstraints = false
-        return textView
-    }()
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-    }
-    
-    func setupView() {
-        addSubview(thumbnailImageView)
-        addSubview(separatorView)
-        addSubview(userProfileImageView)
-        addSubview(titleLabel)
-        addSubview(subtitleTextView)
-        
-     
-        addConstraints(withFormat: "H:|-16-[v0]-16-|", views: thumbnailImageView)
-        addConstraints(withFormat: "H:|-16-[v0(44)]", views: userProfileImageView)
-        
-        // vertical constraint for thumbnail image view and separator view
-        addConstraints(withFormat: "V:|-16-[v0]-8-[v1(44)]-16-[v2(0.5)]|", views: thumbnailImageView, userProfileImageView, separatorView)
-    
-        addConstraints(withFormat: "H:|[v0]|", views: separatorView)
-    
-        // title top constraint 
-        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .top, relatedBy: .equal, toItem: thumbnailImageView, attribute: .bottom, multiplier: 1, constant: 8))
-        
-        // title left constraint
-        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .left, relatedBy: .equal, toItem: userProfileImageView, attribute: .right, multiplier: 1, constant: 8))
-        
-        // title right constraint
-        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .right, relatedBy: .equal, toItem: thumbnailImageView, attribute: .right, multiplier: 1, constant: 0))
-        
-        // title height constraint
-        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .height, relatedBy: .equal, toItem: nil, attribute: .height, multiplier: 1, constant: 20))
-     
-        // subtitle top constraint 
-        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .top, relatedBy: .equal, toItem: titleLabel, attribute: .bottom, multiplier: 1, constant: 8))
-        
-        // subtitle left constraint 
-        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .left, relatedBy: .equal, toItem: userProfileImageView, attribute: .right, multiplier: 1, constant: 8))
-        
-        // subtitle right constraint 
-        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .right, relatedBy: .equal, toItem: titleLabel, attribute: .right, multiplier: 1, constant: 0))
-        
-        //subtitle height constraint
-        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .height, relatedBy: .equal, toItem: nil, attribute: .height, multiplier: 1, constant: 20))
-    }
-}
-
-extension UIView {
-    
-    func addConstraints(withFormat format: String, views: UIView...) {
-        var viewsDictionary = [String: UIView]()
-        for (index, view) in views.enumerated() {
-            let key = "v\(index)"
-            view.translatesAutoresizingMaskIntoConstraints = false
-            viewsDictionary[key] = view
-        }
-        
-        addConstraints(NSLayoutConstraint.constraints(withVisualFormat: format, options: NSLayoutFormatOptions(), metrics: nil, views: viewsDictionary))
-    }
 }
