+++ YoutubeClone/VideoCell.swift
@@ -1,0 +1,104 @@
+//
+//  VideoCell.swift
+//  YoutubeClone
+//
+//  Created by Ky Nguyen on 10/14/16.
+//  Copyright © 2016 Ky Nguyen. All rights reserved.
+//
+
+import UIKit
+
+
+class VideoCell: UICollectionViewCell {
+    
+    let thumbnailImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.image = UIImage(named: "swift-banner")
+        imageView.contentMode = .scaleAspectFill
+        imageView.clipsToBounds = true
+        return imageView
+    }()
+    
+    let separatorView: UIView = {
+        let view = UIView()
+        view.backgroundColor = UIColor.lightGray.withAlphaComponent(0.25)
+        return view
+    }()
+    
+    let userProfileImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.contentMode = .scaleAspectFill
+        imageView.layer.masksToBounds = true
+        imageView.layer.cornerRadius = 22
+        imageView.image = UIImage(named: "swift-banner")
+        return imageView
+    }()
+    
+    let titleLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Lets build that app"
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    let subtitleTextView: UITextView = {
+        let textView = UITextView()
+        textView.translatesAutoresizingMaskIntoConstraints = false
+        textView.text = "Youtube app from Brain (Lets build that app) - 10 views - 1 years ago"
+        textView.textContainerInset = UIEdgeInsets(top: 0, left: -4, bottom: 0, right: 0)
+        textView.textColor = UIColor.lightGray
+        return textView
+    }()
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupView()
+    }
+    
+    func setupView() {
+        addSubview(thumbnailImageView)
+        addSubview(separatorView)
+        addSubview(userProfileImageView)
+        addSubview(titleLabel)
+        addSubview(subtitleTextView)
+        
+        
+        addConstraints(withFormat: "H:|-16-[v0]-16-|", views: thumbnailImageView)
+        addConstraints(withFormat: "H:|-16-[v0(44)]", views: userProfileImageView)
+        
+        // vertical constraint for thumbnail image view and separator view
+        addConstraints(withFormat: "V:|-16-[v0]-8-[v1(44)]-16-[v2(0.5)]|", views: thumbnailImageView, userProfileImageView, separatorView)
+        
+        addConstraints(withFormat: "H:|[v0]|", views: separatorView)
+        
+        // title top constraint
+        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .top, relatedBy: .equal, toItem: thumbnailImageView, attribute: .bottom, multiplier: 1, constant: 8))
+        
+        // title left constraint
+        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .left, relatedBy: .equal, toItem: userProfileImageView, attribute: .right, multiplier: 1, constant: 8))
+        
+        // title right constraint
+        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .right, relatedBy: .equal, toItem: thumbnailImageView, attribute: .right, multiplier: 1, constant: 0))
+        
+        // title height constraint
+        addConstraint(NSLayoutConstraint(item: titleLabel, attribute: .height, relatedBy: .equal, toItem: nil, attribute: .height, multiplier: 1, constant: 20))
+        
+        // subtitle top constraint
+        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .top, relatedBy: .equal, toItem: titleLabel, attribute: .bottom, multiplier: 1, constant: 4))
+        
+        // subtitle left constraint
+        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .left, relatedBy: .equal, toItem: userProfileImageView, attribute: .right, multiplier: 1, constant: 8))
+        
+        // subtitle right constraint
+        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .right, relatedBy: .equal, toItem: titleLabel, attribute: .right, multiplier: 1, constant: 0))
+        
+        //subtitle height constraint
+        addConstraint(NSLayoutConstraint(item: subtitleTextView, attribute: .height, relatedBy: .equal, toItem: nil, attribute: .height, multiplier: 1, constant: 30))
+    }
+}
+
+
